pyproject.toml
[project]
name = "mcp-practice-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "mcp>=1.14.0",
]
